+++ app/src/main/java/com/tripper/db/entities/DaySegment.java
@@ -7,11 +7,11 @@ import androidx.room.PrimaryKey;
 @Entity(tableName = "day_segment")
 public class DaySegment {
     @PrimaryKey(autoGenerate = true)
-    public int id;
+    public long id;
 
     @ColumnInfo(name = "segment")
     public int segment; // 0, 1 , or 2 for morning, afternoon, evening respectively
 
     @ColumnInfo(name= "day_id")
-    public int dayId;
+    public long dayId;
 }
